Fix waiting for visibility status
waitForElementVisible() and waitForElementNotVisible() can fail under
certain circumstances.

For example:
 - If the element already exists and is visible, and later only some
   properties changed, e.g. the element text, waitForElementNotVisible()
   will not detect the fact that the original element no longer exists.
 - If the original element exists and is not visible, then some property
   e.g. the text was changed, and then a new element matching the
   original selector was added and visible, waitForElementVisible()
   would fail.

This is because nightwatch only keeps looking for the element until is
is found. Then it only checks the visibility status of that element. The
fix is to always check both the presence of the item as well as the
visibility.