Emit an error when an element command fails
This can happen, for example, when trying to click() an element which is
covered by another element. In that case we'll get an error from the
WebDriver but so far nightwatch silently ignored such errors.

The caller can handle the error himself by returning a truthy value from
the callback.